--- a/src/poetry/console/commands/source/add.py
+++ b/src/poetry/console/commands/source/add.py
@@ -36,7 +36,6 @@ class SourceAddCommand(Command):
 
     def handle(self) -> int:
         from poetry.factory import Factory
-        from poetry.repositories import RepositoryPool
         from poetry.utils.source import source_to_table
 
         name = self.argument("name")
@@ -84,13 +83,14 @@ class SourceAddCommand(Command):
             self.line(f"Adding source with name <c1>{name}</c1>.")
             sources.append(source_to_table(new_source))
 
+        self.poetry.config.merge(
+            {"sources": {source["name"]: source for source in sources}}
+        )
+
         # ensure new source is valid. eg: invalid name etc.
-        self.poetry._pool = RepositoryPool()
         try:
-            Factory.configure_sources(
-                self.poetry, sources, self.poetry.config, NullIO()
-            )
-            self.poetry.pool.repository(name)
+            pool = Factory.create_pool(self.poetry.config, NullIO())
+            pool.repository(name)
         except ValueError as e:
             self.line_error(
                 f"<error>Failed to validate addition of <c1>{name}</c1>: {e}</error>"