--- a/src/poetry/factory.py
+++ b/src/poetry/factory.py
@@ -29,9 +29,9 @@ if TYPE_CHECKING:
     from poetry.core.packages.package import Package
     from tomlkit.toml_document import TOMLDocument
 
+    from poetry.repositories import RepositoryPool
     from poetry.repositories.legacy_repository import LegacyRepository
     from poetry.utils.dependency_specification import DependencySpec
-
 
 logger = logging.getLogger(__name__)
 
@@ -90,13 +90,15 @@ class Factory(BaseFactory):
         )
 
         # Configuring sources
-        self.configure_sources(
-            poetry,
-            poetry.local_config.get("source", []),
-            config,
-            io,
-            disable_cache=disable_cache,
+        config.merge(
+            {
+                "sources": {
+                    source["name"]: source
+                    for source in poetry.local_config.get("source", [])
+                }
+            }
         )
+        poetry.set_pool(self.create_pool(config, io, disable_cache=disable_cache))
 
         plugin_manager = PluginManager(Plugin.group, disable_plugins=disable_plugins)
         plugin_manager.load_plugins()
@@ -110,23 +112,28 @@ class Factory(BaseFactory):
         return ProjectPackage(name, version)
 
     @classmethod
-    def configure_sources(
+    def create_pool(
         cls,
-        poetry: Poetry,
-        sources: list[dict[str, str]],
         config: Config,
-        io: IO,
+        io: IO | None = None,
         disable_cache: bool = False,
-    ) -> None:
+    ) -> RepositoryPool:
+        from poetry.repositories import RepositoryPool
+
+        if io is None:
+            io = NullIO()
+
         if disable_cache:
             logger.debug("Disabling source caches")
 
-        for source in sources:
+        pool = RepositoryPool()
+
+        for source in config.get("sources", {}).values():
             repository = cls.create_package_source(
                 source, config, disable_cache=disable_cache
             )
-            is_default = bool(source.get("default", False))
-            is_secondary = bool(source.get("secondary", False))
+            is_default = source.get("default", False)
+            is_secondary = source.get("secondary", False)
             if io.is_debug():
                 message = f"Adding repository {repository.name} ({repository.url})"
                 if is_default:
@@ -136,21 +143,23 @@ class Factory(BaseFactory):
 
                 io.write_line(message)
 
-            poetry.pool.add_repository(repository, is_default, secondary=is_secondary)
+            pool.add_repository(repository, is_default, secondary=is_secondary)
 
         # Put PyPI last to prefer private repositories
         # unless we have no default source AND no primary sources
         # (default = false, secondary = false)
-        if poetry.pool.has_default():
+        if pool.has_default():
             if io.is_debug():
                 io.write_line("Deactivating the PyPI repository")
         else:
             from poetry.repositories.pypi_repository import PyPiRepository
 
-            default = not poetry.pool.has_primary_repositories()
-            poetry.pool.add_repository(
+            default = not pool.has_primary_repositories()
+            pool.add_repository(
                 PyPiRepository(disable_cache=disable_cache), default, not default
             )
+
+        return pool
 
     @classmethod
     def create_package_source(