--- a/src/poetry/installation/executor.py
+++ b/src/poetry/installation/executor.py
@@ -16,13 +16,13 @@ from typing import Any
 
 from cleo.io.null_io import NullIO
 from poetry.core.packages.utils.link import Link
-from poetry.core.pyproject.toml import PyProjectTOML
 
 from poetry.installation.chef import Chef
 from poetry.installation.chooser import Chooser
 from poetry.installation.operations import Install
 from poetry.installation.operations import Uninstall
 from poetry.installation.operations import Update
+from poetry.installation.wheel_installer import WheelInstaller
 from poetry.utils._compat import decode
 from poetry.utils.authenticator import Authenticator
 from poetry.utils.env import EnvCommandError
@@ -60,6 +60,8 @@ class Executor:
         self._dry_run = False
         self._enabled = True
         self._verbose = False
+        self._wheel_installer = WheelInstaller(self._env)
+        self._use_wheel_installer = config.get("experimental.wheel-installer", True)
 
         if parallel is None:
             parallel = config.get("installer.parallel", True)
@@ -471,18 +473,21 @@ class Executor:
         message = f"  <fg=blue;options=bold>•</> {op_msg}: <info>Removing...</info>"
         self._write(operation, message)
 
-        return self._remove(operation)
+        return self._remove(operation.package)
 
     def _install(self, operation: Install | Update) -> int:
         package = operation.package
         if package.source_type == "directory":
+            if not self._use_wheel_installer:
+                return self._install_directory_without_wheel_installer(operation)
+
             return self._install_directory(operation)
 
         if package.source_type == "git":
             return self._install_git(operation)
 
         if package.source_type == "file":
-            archive = self._prepare_file(operation)
+            archive = self._prepare_archive(operation)
         elif package.source_type == "url":
             assert package.source_url is not None
             archive = self._download_link(operation, Link(package.source_url))
@@ -495,14 +500,18 @@ class Executor:
             " <info>Installing...</info>"
         )
         self._write(operation, message)
-        return self.pip_install(archive, upgrade=operation.job_type == "update")
+
+        if not self._use_wheel_installer:
+            return self.pip_install(archive, upgrade=operation.job_type == "update")
+
+        self._wheel_installer.install(archive)
+
+        return 0
 
     def _update(self, operation: Install | Update) -> int:
         return self._install(operation)
 
-    def _remove(self, operation: Uninstall) -> int:
-        package = operation.package
-
+    def _remove(self, package: Package) -> int:
         # If we have a VCS package, remove its source directory
         if package.source_type == "git":
             src_dir = self._env.path / "src" / package.name
@@ -517,7 +526,7 @@ class Executor:
 
             raise
 
-    def _prepare_file(self, operation: Install | Update) -> Path:
+    def _prepare_archive(self, operation: Install | Update) -> Path:
         package = operation.package
         operation_message = self.get_operation_message(operation)
 
@@ -532,9 +541,54 @@ class Executor:
         if not Path(package.source_url).is_absolute() and package.root_dir:
             archive = package.root_dir / archive
 
-        return archive
+        return self._chef.prepare(archive)
 
     def _install_directory(self, operation: Install | Update) -> int:
+        package = operation.package
+        operation_message = self.get_operation_message(operation)
+
+        message = (
+            f"  <fg=blue;options=bold>•</> {operation_message}:"
+            " <info>Building...</info>"
+        )
+        self._write(operation, message)
+
+        assert package.source_url is not None
+        if package.root_dir:
+            req = package.root_dir / package.source_url
+        else:
+            req = Path(package.source_url).resolve(strict=False)
+
+        if package.source_subdirectory:
+            req /= package.source_subdirectory
+
+        if package.develop:
+            # Editable installations are currently not supported
+            # for PEP-517 build systems so we defer to pip.
+            # TODO: Remove this workaround once either PEP-660 or PEP-662 is accepted
+            return self.pip_install(req, editable=True)
+
+        archive = self._prepare_archive(operation)
+
+        try:
+            if operation.job_type == "update":
+                # Uninstall first
+                # TODO: Make an uninstaller and find a way to rollback in case
+                # the new package can't be installed
+                assert isinstance(operation, Update)
+                self._remove(operation.initial_package)
+
+            self._wheel_installer.install(archive)
+        finally:
+            archive.unlink()
+
+        return 0
+
+    def _install_directory_without_wheel_installer(
+        self, operation: Install | Update
+    ) -> int:
+        from poetry.core.pyproject.toml import PyProjectTOML
+
         from poetry.factory import Factory
 
         package = operation.package
@@ -644,6 +698,7 @@ class Executor:
     def _download_link(self, operation: Install | Update, link: Link) -> Path:
         package = operation.package
 
+        output_dir = self._chef.get_cache_directory_for_link(link)
         archive = self._chef.get_cached_archive_for_link(link)
         if archive is None:
             # No cached distributions was found, so we download and prepare it
@@ -659,7 +714,16 @@ class Executor:
 
                 raise
 
-        if package.files:
+        if archive.suffix != ".whl":
+            message = (
+                f"  <fg=blue;options=bold>•</> {self.get_operation_message(operation)}:"
+                " <info>Preparing...</info>"
+            )
+            self._write(operation, message)
+
+            archive = self._chef.prepare(archive, output_dir=output_dir)
+
+        if package.files and archive.name in {f["file"] for f in package.files}:
             archive_hash = self._validate_archive_hash(archive, package)
 
             self._hashes[package.name] = archive_hash