--- a/src/poetry/installation/executor.py
+++ b/src/poetry/installation/executor.py
@@ -123,6 +123,9 @@ class Executor:
 
         return self
 
+    def enable_bytecode_compilation(self, enable: bool = True) -> None:
+        self._wheel_installer.enable_bytecode_compilation(enable)
+
     def pip_install(
         self, req: Path, upgrade: bool = False, editable: bool = False
     ) -> int: