… docs/cli.md
… docs/configuration.md
… poetry.lock
… pyproject.toml
… src/poetry/utils/env.py
… tests/conftest.py
… tests/console/commands/self/conftest.py
… tests/console/commands/self/test_update.py
… tests/console/commands/test_add.py
… tests/console/commands/test_config.py
… tests/console/commands/test_install.py
… tests/fixtures/distributions/demo-0.1.0.tar.gz
… tests/fixtures/distributions/demo-0.1.2-py2.py3-none-any.whl
… tests/fixtures/extended_with_no_setup/README.md
… tests/fixtures/extended_with_no_setup/build.py
… tests/fixtures/extended_with_no_setup/extended/extended.c
… tests/fixtures/extended_with_no_setup/pyproject.toml
… tests/installation/test_chef.py
… tests/installation/test_executor.py
… tests/installation/test_wheel_installer.py
… tests/repositories/fixtures/pypi.org/dists/black-21.11b0-py3-none-any.whl
… tests/repositories/fixtures/pypi.org/dists/cleo-1.0.0a5-py3-none-any.whl
… tests/repositories/fixtures/pypi.org/dists/poetry_core-1.5.0-py3-none-any.whl
… tests/repositories/fixtures/pypi.org/dists/pytest-3.5.0-py2.py3-none-any.whl
… tests/repositories/fixtures/pypi.org/dists/pytest-3.5.1-py2.py3-none-any.whl
… tests/repositories/fixtures/pypi.org/json/cleo/1.0.0a5.json
… tests/repositories/fixtures/pypi.org/json/importlib-metadata.json
… tests/repositories/fixtures/pypi.org/json/importlib-metadata/1.7.0.json
… tests/repositories/fixtures/pypi.org/json/poetry-core.json
… tests/repositories/fixtures/pypi.org/json/poetry-core/1.5.0.json