This PR (which is an updated version of #4213) is the first step towards getting rid of subprocess calls to `pip` for package installations and removals.

The wheel building is handled by https://github.com/pypa/build and the wheel installation is handled by https://github.com/pypa/installer (with modifications to be more lenient).

Note that the wheel building is **not** parallelized due to race conditions that I am yet to track down (it seems like the issue comes from `build` when calling the backend hooks). This is not a blocker though.

Overall this speeds up quite a bit the installation times, especially once wheels are cached.

An example with the installation of 115 packages (warm cache and artifacts already downloaded):

Without the new builder/installer: `226.24s user 15.26s system 488% cpu 49.413 total`
With the new builder/installer: `11.52s user 3.08s system 117% cpu 12.452 total`

Note that this can be deactivated by setting `experimental.installer.modern-installation`.

# Pull Request Check List

Resolves: https://github.com/python-poetry/roadmap/issues/1

- [x] Added **tests** for changed code.
- [x] Updated **documentation** for changed code.