@radoering 2022-11-05
Rebased, resolved merge conflicts, swapped the builder and the installer commit (because the builder commit depends on the installer commit) and addressed some of the review comments.

While executing the tests locally, I noticed that e.g. `test_execute_prints_warning_for_yanked_package` does not install in a temp folder anymore but in poetry's own venv. However, I'm not sure if that is an issue of the new `WheelInstaller` or `MockEnv`. It seems that the pip installer respects the temp path via https://github.com/python-poetry/poetry/blob/3a961dfb67ccc214fa101f5c1a187e2ed077673b/src/poetry/utils/pip.py#L34 but the new `WheelInstaller` derives the destination from https://github.com/python-poetry/poetry/blob/3a961dfb67ccc214fa101f5c1a187e2ed077673b/src/poetry/utils/env.py#L1598
which does not return a temp path.