@dimbleby 2022-11-06
I experimented with the arbitrary large-ish install at https://github.com/lincolnloop/python-package-manager-shootout

It looks as though, at least, in that case, the lost parallelism costs more than the faster build buys.

Before:
```bash
$ time poetry install
...
real    1m9.535s
user    5m2.576s
sys     0m40.103s
```
After:
```bash
$ time poetry install
...
real    2m5.472s
user    3m2.383s
sys     1m17.763s
```
so it really has done ~60% of the work - but it has taken nearly twice as long anyway

(both with an empty cache.  With a populated cache the new code is unambiguously faster.  I guess populated cache is more common on a user's own computer, but empty cache is more likely eg in a CI pipeline)