@radoering 2022-11-20
> I experimented with the arbitrary large-ish install at https://github.com/lincolnloop/python-package-manager-shootout

I followed up on this and created a pyproject.toml with the following three dependencies, which only provide a source distribution:

```
maxminddb = "2.2.0"
petname = "2.6"
python-u2flib-server = "5.0.1"
```

Before each `poetry install`, I removed the three wheels from the artifacts cache. Results:

poetry 1.2.2

```
real    0m5.761s
user    0m9.343s
sys     0m0.929s
```

before using lock more precisely

```
real    0m11.785s
user    0m7.965s
sys     0m1.177s
```

after using lock more precisely

```
real    0m6.837s
user    0m9.108s
sys     0m1.899s
```

Although not as fast as with the old installer, still a significant improvement.