@konstin 2022-11-20
The code does compute hashes, but does not compile bytecode, right? I've benchmarked my own wheel installer implementation against pip 22.2.2, poetry 1.2.2 and this branch (`poetry@installer`) using [this sample project](https://github.com/konstin/poc-monotrail/tree/a47b4b953e5042fbc0a5ac86376864cbc0e94702/test-data/poetry/data-science) and [this script](https://github.com/konstin/poc-monotrail/blob/a47b4b953e5042fbc0a5ac86376864cbc0e94702/benchmark.sh) with this branch patched in.

```
test-data/poetry/data-science -E tqdm_feature
pip 22.2.2 from /home/konsti/monotrail/.venv/lib/python3.8/site-packages/pip (python 3.8)
Poetry (version 1.2.2)
Benchmark 1: .venv/bin/pip install -q -r requirements-benchmark.txt
  Time (mean ± σ):     11.868 s ±  0.246 s    [User: 9.395 s, System: 1.324 s]
  Range (min … max):   11.589 s … 12.698 s    20 runs
 
Benchmark 2: poetry install -q --no-root --only main -E tqdm_feature
  Time (mean ± σ):     11.784 s ±  0.410 s    [User: 23.019 s, System: 3.801 s]
  Range (min … max):   10.058 s … 12.069 s    20 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Benchmark 3: poetry@installer install -q --no-root --only main -E tqdm_feature
  Time (mean ± σ):      3.680 s ±  0.059 s    [User: 3.553 s, System: 0.949 s]
  Range (min … max):    3.637 s …  3.899 s    20 runs
 
Benchmark 4: /home/konsti/monotrail/target/release/monotrail poetry-install -E tqdm_feature
  Time (mean ± σ):      4.885 s ±  0.122 s    [User: 12.715 s, System: 2.162 s]
  Range (min … max):    4.676 s …  5.082 s    20 runs
 
Summary
  'poetry@installer install -q --no-root --only main -E tqdm_feature' ran
    1.33 ± 0.04 times faster than '/home/konsti/monotrail/target/release/monotrail poetry-install -E tqdm_feature'
    3.20 ± 0.12 times faster than 'poetry install -q --no-root --only main -E tqdm_feature'
    3.22 ± 0.08 times faster than '.venv/bin/pip install -q -r requirements-benchmark.txt'
```

If i also skip bytecode compilation on pip and mine i got those numbers for a short run (3 repetitions):

```
test-data/poetry/data-science -E tqdm_feature
pip 22.2.2 from /home/konsti/monotrail/.venv/lib/python3.8/site-packages/pip (python 3.8)
Poetry (version 1.2.2)
Benchmark 1: .venv/bin/pip install --no-compile -q -r requirements-benchmark.txt
  Time (mean ± σ):      6.238 s ±  0.172 s    [User: 4.419 s, System: 0.899 s]
  Range (min … max):    6.070 s …  6.413 s    3 runs
 
Benchmark 2: poetry install -q --no-root --only main -E tqdm_feature
  Time (mean ± σ):     10.266 s ±  0.008 s    [User: 18.091 s, System: 2.818 s]
  Range (min … max):   10.259 s … 10.275 s    3 runs
 
Benchmark 3: poetry@installer install -q --no-root --only main -E tqdm_feature
  Time (mean ± σ):      3.560 s ±  0.081 s    [User: 3.396 s, System: 0.901 s]
  Range (min … max):    3.495 s …  3.650 s    3 runs
 
Benchmark 4: /home/konsti/monotrail/target/release/monotrail poetry-install -E tqdm_feature
  Time (mean ± σ):      5.648 s ±  1.659 s    [User: 14.845 s, System: 2.623 s]
  Range (min … max):    3.872 s …  7.157 s    3 runs
 
Benchmark 5: /home/konsti/monotrail/target/release/monotrail poetry-install --no-compile -E tqdm_feature
  Time (mean ± σ):      2.070 s ±  0.226 s    [User: 3.513 s, System: 1.018 s]
  Range (min … max):    1.935 s …  2.330 s    3 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Summary
  '/home/konsti/monotrail/target/release/monotrail poetry-install --no-compile -E tqdm_feature' ran
    1.72 ± 0.19 times faster than 'poetry@installer install -q --no-root --only main -E tqdm_feature'
    2.73 ± 0.85 times faster than '/home/konsti/monotrail/target/release/monotrail poetry-install -E tqdm_feature'
    3.01 ± 0.34 times faster than '.venv/bin/pip install --no-compile -q -r requirements-benchmark.txt'
    4.96 ± 0.54 times faster than 'poetry install -q --no-root --only main -E tqdm_feature'
```