@radoering 2022-12-04
> The code does compute hashes, but does not compile bytecode, right?

Good catch. It seems we did compare apples to oranges so far (at least to some degree). If I'm not mistaken, the "old" installer did compile bytecode but this one doesn't. I wasn't aware of that and I assume it's not intentional (but actually I don't know).