@Secrus 2022-12-07
Two side-notes about dependencies: 
- `pep517` is now known as [`pyproject-hooks`](https://github.com/pypa/pyproject-hooks) and neither the old nor new version is defined in `pyproject.toml` as a top-level dependency (used in chef implementation)
- `installer` just got a new version [0.6.0](https://pypi.org/project/installer), which exports typing information so `type: ignore` comments should no longer be needed