@radoering 2022-12-11
Rebased on current master and
* updated build and installer
* removed `type: ignore` comments no longer needed for installer
* replaced `pep517` usage by `pyproject-hooks` and added it as explicit dependency
* fixed version in INSTALLER file
* added tests for `WheelInstaller`
* added `--compile` option to `poetry install` (@konstin We decided to make bytecode compilation opt-in.)